Extract execute_instruction() to seed new runtime module
Fixes #1528
