+++ src/lib.rs
@@ -65,6 +65,7 @@ pub mod retransmit_stage;
 pub mod rpc;
 pub mod rpc_pubsub;
 pub mod rpc_request;
+pub mod runtime;
 pub mod service;
 pub mod signature;
 pub mod sigverify;
